Completed new payment page with PayPal sanbox intergration
Added an example of a transaction with a image of the car, information about the car, map details of where to pick up and drop off the car as well as time. Also a summary of charges and a button which redirects to PayPal's sign in page (Credit card can also be used at this page as well).