added session for logging in and out
added session for logging in and out, fixed navbar, session determines what to display on the navbar. please use this new navbar and header. I started working on displaying the correct data from the database for the my account page, only issue im having trouble with is it's showing the wrong data so this isnt final, once I can fix this I can go and change all the hardcoded values to the correct ones. also placed the background image we're going to use in the files.